python/fastapi-graphql-with-authz/pyproject.toml
[tool.poetry]
name = "fastapi-graphql-with-authz"
version = "0.1.0"
description = "FastAPI에 인가시스템 적용기"
authors = [
    "Hannal <kay@hannal.net>",
]
packages = [
    { include='fastapi_authz', from='.' },
    { include='apps', from='.' },
]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "~0.23.0"
pydantic = {extras = ["email"], version = "~1.9.1"}
fastapi = "~0.79.0"
uvicorn = { extras = ["standard"], version = "~0.17.6"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.122.1"}
python-multipart = "^0.0.5"
ujson = "~5.4.0"
email-validator = "^1.2.1"
requests = "~2.28.1"
PyYAML = "~6.0"
SQLAlchemy-Utils = "^0.38.3"
SQLAlchemy = {extras = ["postgresql_asyncpg"], version = "^1.4.41"}
alembic = {extras = ["tz"], version = "^1.8.1"}
greenlet = "^1.1.3"
typer = {extras = ["all"], version = "^0.6.1"}
aiosqlite = "^0.17.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
factory-boy = "~3.2.1"
types-ujson = "~5.2.0"
asgi-lifespan = "^1.0.1"
black = "^22.8.0"
isort = "^5.10.1"
pytest = "^7.1.3"
pytest-mock = "^3.8.2"
pytest-cov = "^3.0.0"
pytest-env = "^0.6.2"
pytest-asyncio = "^0.19.0"
pytest-dependency = "^0.5.1"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.py4$'
exclude = '/(\.venv|venv|\.env|env|migrations|stubs)/'

[tool.pytest.ini_options]
testpaths = "tests"
python_files = ["tests.py", "test_*.py"]
addopts = '''
  --strict-markers
  --tb=short
  --no-cov-on-fail
  --asyncio-mode=auto
'''
env = []
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:etcd3.*:",
]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
omit = ["*/tests/*", "*/conftest.py", "*/app.py", "*/main.py"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "def __str__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "if typing.TYPE_CHECKING:"
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.coverage.json]
output = "coverage.json"
